requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.62.1
kiwisolver==1.5.0
litemapy==0.11.0b0
matplotlib==3.10.8
nbtlib==2.0.4
numpy==2.4.4
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pycolmap==4.0.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
seaborn==0.13.2
six==1.17.0
typing_extensions==4.15.0
opencv-python
open3d>=0.17.0
scipy
